Metodo OrderCreateDto

+++ apiCliente/src/main/java/ws/dto/raspay/OrderDto.java
@@ -1,5 +1,7 @@
 package ws.dto.raspay;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class OrderDto {
 	private String id;
 	private String customerId;
-	private Long dicount;
+	private BigDecimal dicount;
 	private String productAcronym;
 	
 }
